Refactoring Create- and EditQuestionnaireViewModels into  and ~Question into two separate ViewModels. Changed variable names for consistency and legibility. Made further necessary changes surrounding the ViewModelLocator
