+++ FSBeheer/FSBeheer/View/EditQuestionnaireView.xaml
@@ -7,7 +7,7 @@
         xmlns:local="clr-namespace:FSBeheer.View"
         mc:Ignorable="d"
         Title="Vragenlijst Aanpassen" Height="500" Width="800"
-        DataContext="{Binding EditQuestionnaireViewModel, Source={StaticResource Locator}}">
+        DataContext="{Binding EditQuestionnaire, Source={StaticResource Locator}}">
     <StackPanel>
         <StackPanel  Margin="0,10,0,0" Width="500">
             <TextBlock>Inspectienaam:</TextBlock>
